+++ src/main/java/subway/application/PathService.java
@@ -6,11 +6,11 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.common.util.PathFinder;
 import subway.dao.PathDao;
 import subway.dao.SectionDao;
 import subway.dao.StationDao;
 import subway.domain.Path;
+import subway.domain.PathFinder;
 import subway.domain.Station;
 import subway.dto.PathFinderResult;
 import subway.dto.PathResponse;
